+++ CHANGELOG.md
@@ -1068,3 +1068,4 @@ Tab switching changed from onTap: (_) => setState(() {}) which rebuilt on every 
 - Wrapped the recent foods list in a ConstrainedBox with a max height and SingleChildScrollView to make the tab scrollable and prevent pixel overflow errors
 - Moved the attribution text out of the buildSearchTab and buildBarcodeTab methods and moving them to the body's main column (above the recent foods section) so that expanding the recent foods tab did not push it down
 - Edited textDropShadow to make the shadows of text more subtle
+- Added a NotificationListener to listen to user scrolls, and a GestureDetector to detect user clicks, to automatically collapse the recent logs tab if the user scrolls / clicks on their logged foods under it, improving UX